Home Page Outline + Stylesheet updates
I completed the general setup of the Home screen, so now all that needs to visually be done is adding icons and progress bars.

Note: Changes have been made to Stylesheet to organize the text as well as hold new fields I needed to use for my frames on the Home screen.